rustdoc: Extract footnote logic into it's own module.
